+++ backup/src/routes/(editor)/editor/views/code.svelte
@@ -1,7 +1,0 @@
-<script>
-    import { globalComponentCollectionStore } from "../../../globals/globalstores.js";
-</script>
-
-<div class="container-fluid">
-    <pre>{JSON.stringify($globalComponentCollectionStore, null, 2)}</pre>
-</div>