+++ backup/src/routes/(modules)/widgets/row/rowoptions.svelte
@@ -1,1 +1,0 @@
-<div>Row Options</div>